requirements.txt
pandas
numpy
python-dotenv
pypdf
python-docx
sentence-transformers
faiss-cpu
langchain
streamlit
